[Pal] Deduplicate _DkGetCPUInfo implementations
And also move them to a proper, arch-specific, but not PAL-specific
directory.
