+++ Pal/include/host/Linux-common/linux_utils.h
@@ -8,9 +8,6 @@
 #include <stdint.h>
 #include <stdnoreturn.h>
 
-double get_bogomips_from_cpuinfo_buf(const char* buf);
-double sanitize_bogomips_value(double);
-
 char* get_main_exec_path(void);
 
 /* Usable only for blocking FDs */
