+++ Pal/src/host/Linux-common/bogomips.c
@@ -1,5 +1,7 @@
 #include "api.h"
 #include "linux_utils.h"
+#include "pal_internal.h"
+#include "topo_info.h"
 
 /* This version is too dumb to be shared by the whole repository and should be removed once we get
  * a proper stdlib (like musl). */
@@ -42,16 +44,29 @@ static char* find_entry_in_cpuinfo(const char* cpuinfo, const char* word) {
     return NULL;
 }
 
-double get_bogomips_from_cpuinfo_buf(const char* buf) {
+static double get_bogomips_from_cpuinfo_buf(const char* buf) {
     char* start = find_entry_in_cpuinfo(buf, "bogomips");
     if (!start)
         return 0.0;
     return proc_cpuinfo_atod(start);
 }
 
-double sanitize_bogomips_value(double v) {
+static double sanitize_bogomips_value(double v) {
     if (!__builtin_isnormal(v) || v < 0.0) {
         return 0.0;
     }
     return v;
+}
+
+double _DkGetBogomips(void) {
+    /* Although the whole file might not fit in this size, the first cpu description should. */
+    char buf[2048];
+    ssize_t len;
+
+    len = read_file_buffer("/proc/cpuinfo", buf, sizeof(buf) - 1);
+    if (len < 0)
+        return 0.0;
+    buf[len] = 0;
+
+    return sanitize_bogomips_value(get_bogomips_from_cpuinfo_buf(buf));
 }
