+++ Pal/src/meson.build
@@ -19,4 +19,11 @@ pal_common_sources = files(
     'slab.c',
 )
 
+pal_common_sources_arch = [
+    'pal_cpuid.c',
+]
+foreach src : pal_common_sources_arch
+    pal_common_sources += files(join_paths('arch', host_machine.cpu_family(), src))
+endforeach
+
 subdir('host')
